'stbt optimise': Optimises a test script generated by 'stbt record'
Performs source-to-source transformations on the specified test script.

Currently only removes calls to wait_for_match where the image passed to
wait_for_match cannot be found on disk. This is to ease the common
workflow of recording a script with 'stbt record', then deleting the
screenshots that aren't of interest (and cropping the rest to the region
of interest for wait_for_match).

To write your own transformations see ast.NodeTransformer in the python
standard library:
http://docs.python.org/library/ast.html#ast.NodeTransformer

Surprisingly, the python standard library doesn't provide any way to
write the ast (abstract syntax tree) back to python source code. So
we use codegen.py from
http://dev.pocoo.org/hg/sandbox/file/98ce1ce17c7c/ast/codegen.py
(it has a BSD license).

According to
http://stackoverflow.com/questions/768634/python-parse-a-py-file-read-the-ast-modify-it-then-write-back-the-modified#comment-5278782
"codegen is largely out-of-date and has a few bugs". For possible
alternatives see
http://stackoverflow.com/questions/768634/python-parse-a-py-file-read-the-ast-modify-it-then-write-back-the-modified#answer-769199