Introduce concept of self-contained stbt directories (nodes)
stbt applies a bunch of tests to a set-top box using some test hardware
(e.g. IR blaster, video capture) set up in a specific manner.  It produces
logs as output from these test runs.  A single PC may have multiple set-top
boxes attached to it at one time or may be set-up to test multiple
different set-top-boxes which require a different configuration and for
which test results, etc. should be generated and kept seperately.

This commit introduces the concept of a stbt "node".  This is a directory
under which configuration, logs and anything else in the future that might
be required can be stored and versioned.  It is to some extent conceptually
inspired by git repositories.

It is intended to have the following structure:

    log/
        results/
    stbt.conf
    test-scripts/
        .git/
        tests/
            <test name>/
                test.py
                <my image>.png
                ...
        library/
            <my helper functions>.py
            ...

and it is recommend that it should be versioned with git but seperately
from the test-scripts themselves.

A stbt node directory is identified by the presence of stbt.conf which
will override configuration set as system and user defaults.

This commit lays some groundwork for this setup by updating get_config to
search for a stbt node upwards from the current working directory.

TODO:
 * Come up with a better name than "nodes"
 * Add documentation (possibly copied from this commit message)