Add stbt-init for easy creation of stbt "nodes"
Adds a tool to create a pristine stbt node[1] to make creating stbt nodes
easier and more standardized.  In the future it would probably be useful
to add an interactive mode to this tool such that it will guide you through
detecting and adding video-capture devices, ir blasters, etc.

TODO:
* May as well make creating a pristine node a git clone of something that
  comes with stbt if we're producing near identical git repositories anyway
* Investigate actual use-cases
* Add documentation to man page

[1] See the previous commit for more information.