RPM spec file: Don't hard code version/paths
Generate RPM spec file with:

    make prefix=/usr extra/stb-tester.spec

This allows installation directories and version to be automatically
substituted into the rpm spec file.  This change was motivated by future
work which will install a GStreamer plugin.  This is required as the
location into which the GStreamer plugin will be installed depends on the
output of

    pkg-config --variable=pluginsdir gstreamer-1.0

which cannot be reliably known ahead-of-time.