Refactor: Only assign to self.start_pipeline once the pipeline is ready
I prefer this more transactional style of coding because it means that
`self.source_pipeline` never refers to a half-constructed pipeline, even
in the presence of exception.  This makes it easier to verify in-variants.

This is an incidental change made in the process of creating the next
commit fixing latency bugs.