Refactor: _tesseract: Allow passing a `Gst.Sample` as frame
`match_text`, which will be added in the next commit, will return the
timestamp of the frames it analyses.  To do this it will need the extra
metadata attached to `Gst.Sample` that is not available when dealing with
images in `numpy.ndarray` format.  To support this this commit introduces
the ability to pass a `Gst.Sample` to `_tesseract` itself.  It has the
pleasant side-effect of reducing copying.