Makefile: Add docker-build target
Creates a docker image from the current status in git.  This is useful if
you want a docker image of a development version of stb-tester, rather than
using one of the "[trusted builds]".

When building a Dockerfile docker keeps a cache of intermediate stages.  It
decides if it needs to re-run a command from the Dockerfile if the command
changes, or in the case of the ADD command, if the files that are being
added to the container change.  These changes include file metadata such as
mtime and atime.  I use `touch` here to normalise the timestamps of all the
files from the tarball to the UNIX epoch.  This avoids defeating docker's
caching so we don't have to download all the package dependencies every
time we type `make docker-build` as long as the list of dependencies don't
change.

[trusted builds]: http://blog.docker.com/2013/11/introducing-trusted-builds/