extra/service: Generate png version of architecture diagram
Generate a png from the svg, and commit it so that it shows in the
github web interface's view of the readme file.

I had to edit the svg in inkscape because a couple of the arrows were
pointing in weird directions in the output png, even though they looked
fine in the svg in firefox.