Fix template image name in MatchResult string representation
I think this is a bug introduced in df23d1d1. The debug output from
`wait_for_match` etc. always said "<Custom image>"; it never had the
template filename.

It'd be great if we could just attach the filename to the numpy array,
so we could just set `MatchResult.image = the_numpy_array`, and not have
to worry about keeping track of where "image" refers to a filename and
where it refers to a numpy array. Alternately we could set
`MatchResult.image = AnnotatedTemplate(...)` but then we'd have to
expose `AnnotatedTemplate` to the user.