Makefile: Add docker-build target
Creates a docker image from the current status in git.

This also adds the docker-publish target which will publish docker images
to the [Docker Hub] to allow other users to make use of them without having
to build stb-tester itself.

When building a Dockerfile docker keeps a cache of intermediate stages.  It
decides if it needs to re-run a command from the Dockerfile if the command
changes, or in the case of the ADD command, if the files that are being
added to the container change.  These changes include file metadata such as
mtime and atime.  I use `touch` here to normalise the timestamps of all the
files from the tarball to the UNIX epoch.  This avoids defeating docker's
caching so we don't have to download all the package dependencies every
time we type `make docker-build` as long as the list of dependencies don't
change.

There's support here for multiple different Dockerfiles at the root,
although for now we only have the one.

[Docker Hub]: https://hub.docker.com/