Display: Don't hold lock while iterating frames
The purpose of the lock was to allow time for the `Display` to draw on the
frame before passing it to the sink pipeline.  This is a difficult design
to make usable from multiple threads.  Instead we buffer up 1s worth of
frames in `SinkPipeline`, so the test script has 1s to draw on the frame
before it's too late.

This does cause a change in behaviour.  Instead of video running at 1fps
when we're matching at 1fps it will run a full speed.  The annotations will
appear and disappear very quickly.