Display: Allow multiple threads to call `get_frame()` and `frames()`
Previously `get_frame()` would have a side-effect that we'd never get the
same frame twice.  Instead it would block waiting for a new frame.  This is
useful behaviour for simple loops like:

    wait_until(lambda: match("example.png"))

but it's less desirable for:

    wait_until(lambda: match("example.png") or match("alternative.png"))

as waiting for a different frame to match "alternative.png" is unnecessary
and slows the process down.

It goes particularly wrong for cases like:

    for x, y in izip(detect_motion(),
                     iter(lambda: stbt.match("error-dialog.png"))):
        ...

Which frames will be analysed by `detect_motion` and which matched against?
It's unclear.

With this change a call to `get_frame()` will return immediately with the
latest frame and multiple `stbt.frames()` iterators can be used
concurrently (even from separate threads).  This will make it easier to
implement some more complex performance measurement tests.

The effects of this change could be rather subtle.  I would expect there to
be some loops out there that become busy-waits.  Whether that will actually
cause problems or not is a seperate question.