Fix `get_config("a", "b", type_=bool)` for falsey values
Previously if your config file said:

    [section]
    key = false

You would find that `get_config("section", "key", type_=bool) == True`.
This is surprising behaviour as you've explicitly said you're expecting
a bool.  This fixes that issue.

Techinically this is a breaking change, but I don't think it's likely to
hit anyone in practice.