sqdiff: Add numba implementation for comparison purposes
Our C implementation usually outperforms the numba implementation by a
factor of 4-8x, but sometimes it is slower than numba.  This seems to be
caused by fixed overheads as it's only ever faster on very small templates
when the fixed-costs dominate.  For example:

    All times in ms             numpy		 numba
    type    	numpy	numba	C	speedup	speedup	size	alignment

    template 	0.01	0.00	0.02	0.77	0.28	2 x 277 	6
    with mask	0.01	0.00	0.01	0.90	0.29	2 x 277 	6
    unmasked 	0.01	0.00	0.01	0.92	0.19	2 x 277 	6

In this instance the comparisons are taking ~5-10us.

On larger templates you see the usual where the C implementation is faster:

    All times in ms             numpy		 numba
    type    	numpy	numba	C	speedup	speedup	size	alignment

    template 	13.68	1.01	0.13	101.87	7.52	1055 x 418 	3
    with mask	13.71	1.01	0.14	99.06	7.30	1055 x 418 	3
    unmasked 	13.62	1.01	0.13	101.52	7.53	1055 x 418 	3

In this case the C implementation is taking ~130us.